+++ extensions/session/server/index.js
@@ -92,6 +92,7 @@ var start = function (u, p, s, d, req, response, status) {
         req.session.subUser = data.subuser;
         req.session.screenName = data.screen_name;
         req.session.parentDb = data.parentdb;
+        req.session.custom = data.custom;
 
         console.log("Session started");
 
@@ -102,7 +103,8 @@ var start = function (u, p, s, d, req, response, status) {
             email: data.email,
             api_key: data.api_key,
             parentdb: data.parentdb,
-            subuser: data.subUser
+            subuser: data.subUser,
+            custom: data.custom
         };
 
         if (autoLogin) {
@@ -157,7 +159,8 @@ router.get('/api/session/status', function (req, response) {
                 authenticated: !!req.session.gc2SessionId,
                 screen_name: req.session.gc2UserName,
                 email: req.session.gc2Email,
-                subuser: req.session.subUser
+                subuser: req.session.subUser,
+                custom: req.session.custom
             }
         });
     }